+++ examples/text_layouter_inline_boxes.rb
@@ -61,6 +61,6 @@ end
 layouter = TextLayouter.new
 layouter.style.align = :justify
 layouter.style.line_spacing(:proportional, 1.5)
-layouter.fit(items, width: 500, height: 700).draw(doc.pages.add.canvas, 50, 800)
+layouter.fit(items, 500, 700).draw(doc.pages.add.canvas, 50, 800)
 
 doc.write("text_layouter_inline_boxes.pdf")