+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Changed
 
+* [HexaPDF::Layout::TextLayouter#fit] to allow fitting text into arbitrarily
+  shaped areas
 * [HexaPDF::Layout::TextLayouter] to return a new
   [HexaPDF::Layout::TextLayouter::Result] structure when `#fit` is called that
   includes the `#draw` method